+++ Runtime/Values/KeyPressTypeEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace Cobilas.Unity.Management.InputManager {
+    public enum KeyPressType : byte {
+        Press = 0,
+        PressDown = 1,
+        PressUp = 2,
+        AnyPress = 3,
+    }
+}