+++ Test/Runtime/InputManagerTypeEnum.cs
@@ -1,7 +1,0 @@
-﻿namespace Cobilas.Unity.Management.InputManager {
-    public enum InputManagerType : byte {
-        KeyboardCommand = 0,
-        MouseCommand = 1,
-        MixedCommand = 2
-    }
-}