+++ Test/Runtime/KeyPressTypeEnum.cs
@@ -1,8 +1,0 @@
-﻿namespace Cobilas.Unity.Management.InputManager {
-    public enum KeyPressType : byte {
-        Press = 0,
-        PressDown = 1,
-        PressUp = 2,
-        AnyPress = 3
-    }
-}